// src/api/sqlmodule/model/schemaColumnResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { SchemaParameterResponse } from './schemaParameterResponse';

export interface SchemaColumnResponse {
  id: string;
  name: string;
  physicalTypeId: string;
  physicalTypeName: string;
  isPrimaryKey: boolean;
  isRequired: boolean;
  sortOrder: number;
  parameters: SchemaParameterResponse[];
}