// src/api/sqlmodule/model/tableRowChange.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { TableRowChangeCells } from './tableRowChangeCells';
import type { TableRowChangeOperation } from './tableRowChangeOperation';

export interface TableRowChange {
  operation?: TableRowChangeOperation;
  /** @nullable */
  id?: string | null;
  /** @nullable */
  tempId?: string | null;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  sortOrder?: number | null;
  /** @nullable */
  cells?: TableRowChangeCells;
}