// src/api/sqlmodule/model/tableRowResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */
import type { TableRowResponseCells } from './tableRowResponseCells';

export interface TableRowResponse {
  id: string;
  version: string;
  sortOrder: number;
  cells: TableRowResponseCells;
}