// src/api/sqlmodule/model/tableRowOperation.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SQLModule API
 * API SQL-тренажёра (модуль Scoodle).
 * OpenAPI spec version: v1
 */

export type TableRowOperation = typeof TableRowOperation[keyof typeof TableRowOperation];


export const TableRowOperation = {
  Create: 'Create',
  Update: 'Update',
  Delete: 'Delete',
} as const;